runtime/src/storage: Move untrusted in-memory key/value storage
The untrusted in-memory key/value storage is now accessible
to all tests.
