+++ .changelog/5561.internal.md
@@ -1,0 +1,3 @@
+runtime/src/storage: Move untrusted in-memory key/value storage
+
+The untrusted in-memory key/value storage is now accessible to all tests.
